+++ changelog.md
@@ -2,6 +2,7 @@
 * [Feature] Increase race control alerts display time from 10 to 15 seconds.
 * [Feature] Added tool update checker to notify you if a new version is available in the forum.
 * [Feature] Increase trigger time to scroll race results from 10 seconds to 25 seconds.
+* [Bug Fix] Prevent the Driver Info widget overlapping the qualifying results at session end.
 
 ### Version 1.0.0
 * [Feature] Added global LESS variable definitions for themes.